env: disable aux/timesync in Plan 9 image
TestVariousDeadlines is regularly failing on
the plan9/386 builder running on GCE.

Richard Miller suspects that the failure
may be related to an issue in time measurement,
due to the clock going backward and forward.

A work around would be to disable auth/timesync
to prevent prevent the time from jumping around.

This changes fixes the Plan 9 image by disabling
aux/timesync before starting the buildlet.

Fixes golang/go#26945.

Change-Id: Ibe80cdcd1360cd9e40da70c2c454b6ba2eb44ce9
Reviewed-on: https://go-review.googlesource.com/129098
Reviewed-by: Richard Miller <millerresearch@gmail.com>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>