maintner/maintnerd, cmd/gerritbot, cmd/gitmirror: add golang/dl to whitelist
In golang/go#23223, we've created a new subrepository:

	https://go.googlesource.com/dl

Unlike other subrepositories, there isn't a mirror on GitHub yet:

	https://github.com/golang/dl (404 at this moment)

We want to include it in the whitelist for mirroring to GitHub,
including mirroring its pull requests to Gerrit. The motivation is to
make this subrepository consistent with all others and prevent people
from running into an unexpected situation when some repos are mirrored
but some aren't.

We don't actually expect much contribution activity to the dl subrepo,
because it's very narrow in scope and mostly generated code.

Updates golang/go#26949.

Change-Id: Id995bb3d97b8eb328e6c6bba6714bc74fbefb942
Reviewed-on: https://go-review.googlesource.com/133896
Reviewed-by: Andrew Bonventre <andybons@golang.org>