x/build/devapp: put performance issues in their own grouping
Separate performance issues (those with the "Performance" label) from
the rest of the issues on the release dashboard. Performance issues
tend to be those which are not critical to fix for any particular
release, so separating them and putting them later in order helps to
deprioritize them.

Kind of a hack, but it works.

Change-Id: I6fa0f85629504c298691b492e50e03c817c62a1b
Reviewed-on: https://go-review.googlesource.com/c/145937
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>