cmd/updatecontrib: add updatecontrib command
This change adds a updatecontrib command that is used to update
the CONTRIBUTORS file in the Go repository.

It's based on a previously-internal updateac tool that updated both
CONTRIBUTORS and AUTHORS files. That program was modified to implement
the new behavior for golang.org/issue/12042, now that we're no longer
updating the AUTHORS file.

Specifically, all code to check CLAs and update AUTHORS file was
removed. Instead of relying on an external sortac binary from
GOROOT/misc/sortac, its code is incorporated directly here. That
means the misc/sortac binary can be deleted after this is merged.

Updates golang/go#12042

Change-Id: I45ae247c4e5b558e8a11cdaf720b60c2cf8fc876
Reviewed-on: https://go-review.googlesource.com/c/157238
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>