cmd/tip: start building website from x/website repository
As part of golang/go#29206, the golang.org website is being moved to
from x/tools to x/website repository. This change updates the tip
command to start deploying it from its new location.

The new website is using Go modules for reproducible builds. This
simplifies the code, because we no longer need to manually fetch
any dependent repositories (such as x/net).

Update instances of "godoc" (the binary that was previously used to
host the website) to "golangorg", the new binary name in x/website.

Fixes golang/go#30232
Updates golang/go#29206

Change-Id: I428eddf95a7ed6a43a25138ffeb919e79f42f9d2
Reviewed-on: https://go-review.googlesource.com/c/162908
Reviewed-by: Andrew Bonventre <andybons@golang.org>