cmd/tip: add staging deployment configuration
We want to be able to test cmd/tip changes on staging before pushing
to production. This change adds configuration files to be used for
the staging environment.

Mirror the production environment, but use a staging-specific load
balancer IP, domain, and autocert GCS bucket (golang-tip-dev-autocert).

Also rename the existing production configuration files to mirror the
pattern in other directories. This is especially important now that
there are both production and staging configuration files.

Updates golang/go#30232

Change-Id: I11c8cf643654d2c5bfa4add50f9ac8ca6771606b
Reviewed-on: https://go-review.googlesource.com/c/162914
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>