cmd/gopherbot: don't return early on error in doTasks
The purpose of this change is to make gopherbot less fallible to
sleep-error loops.

When running in daemon mode in production, doTasks is responsible
for running all gopherbot tasks. It does so in sequence. If one of
the tasks happens to fail for some reason, doTasks returns right
away and reports the error.

If the failing task happens to be the last task in the sequence,
gopherbot will continue to be mostly functional and perform the
problem-free tasks, but sleep for 30 seconds between loops.
If the failing task happens to be first in sequence, all other
tasks don't get a chance to run, bringing gopherbot to a halt.

A more fair and consistent behavior is to make doTasks run through
all the tasks even if some fail, then report failures at the end.

Fixes golang/go#30182 (second half)

Change-Id: If7e1b0aa0c56e7f4ec35239ca146a56b6eb66ef3
Reviewed-on: https://go-review.googlesource.com/c/164157
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>