cmd/gomote: add -firewall flag to run, defaulting to off
While debugging golang/go#25386 I found that the firewall defaulting
to on made debugging modules hard. So make gomote default to
no outbound firewall and make it opt-in for people who
want to reproduce the builder more.

Also in this CL: start to use server's builder config, not local state
(follow up to CL 169678 which was incomplete). It's still incomplete
in this CL, but there's now a panic with more useful information to
users telling them to update their binary.

Updates golang/go#30929

Change-Id: I17bded71919af1e7a9181866a1349eb72da40051
Reviewed-on: https://go-review.googlesource.com/c/build/+/170397
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>