devapp/owners: remove catchall entry for the 'go' repo
The owners table is consumed by gopherbot to assign reviewers and
escalate triaged issues.

Gopherbot has its own fallback behavior for unowned paths, which
should generally be to escalate to whichever member of the Go team is
on triage duty (or to escalate to the Go Open Source Project subteam
in general to update the owners table).

Due to the explicit entry for the root of the 'go' repository, when
gopherbot encounters an issue or CL for any standard-library package
without an owner, it will instead escalate to rsc, ianlancetaylor, and
bradfitz. That doesn't seem like a great use of their time, and also
masks the missing entry in the owners table.

This CL removes that top-level entry, so that unknown paths within the
Go repo will appear to gopherbot as such.

Updates golang/go#27586

Change-Id: I469028f7ec5f4b357ae33bca125da2323fc3e0d2
Reviewed-on: https://go-review.googlesource.com/c/build/+/171358
Run-TryBot: Bryan C. Mills <bcmills@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>