cmd/scaleway: set page size to 100 when listing servers
According to Scaleway API documentation¹, most endpoints are paginated.
Use the maximum page size of 100 when listing servers. The default page
size was 50, which meant some servers weren't included in the response.

We currently have just over 50 servers, so this is sufficient for now,
but we'll need to do proper pagination when we need to handle over 100.
At that point, need to decide if it's better to implement it ad-hoc here
or start using an existing Scaleway API client. For now, add a TODO and
check that we haven't exceeded the page size, to avoid silent problems.

¹ https://developer.scaleway.com/#header-pagination

Updates golang/go#32229

Change-Id: I254671e464e88017eee2e49e382c686c52fd8fbd
Reviewed-on: https://go-review.googlesource.com/c/build/+/179182
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>