app/appengine: correctly align subheading2 in subrepo tables
Currently, the subheading2 cells in the subrepo tables are shifted by
one to the left due to the additional cell for the Go commit. There are
4 fixed <th> cells while there are 5 <td> cells (subrepo commit, Go
commit, user, time and description).

Make the first <th> have a colspan of 2 in case of the subrepo table so
the subheading2 cells align correctly.

Change-Id: I1ca7157cecf6530cff5883154f367684b0695081
Reviewed-on: https://go-review.googlesource.com/c/build/+/183217
Run-TryBot: Tobias Klauser <tobias.klauser@gmail.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>