cmd/releasebot: don't mention release status issue in security release mode
There is no public release status issue in security release mode,
because we want to avoid the risk of leaking sensitive test failures.

Don't write the release status issue title to stdout, because that
makes it seem like there might be a public issue, which can be
worrisome to see when doing a security release.

Change-Id: Ice9b3fe1067a8d68cf8d22cd944a3b92a3abb606
Reviewed-on: https://go-review.googlesource.com/c/build/+/190801
Reviewed-by: Alexander Rakoczy <alex@golang.org>