maintner: add option to prefix gcsLogSegments
Can optionally specify a prefix for gcsLogSegments when constructing
a GCSLog. This allows for a single bucket to be used with different
mutation logs scoped to "folders" by the prefix.

It should be noted that the notion of folders is arbitrary as in GCS, objects
are stored in a flat structure but can appear nested when combined with
unix-like paths.

Change-Id: Ia4b3c9d3b9dec7f1cca731f8546d80a64b1889b0
Reviewed-on: https://go-review.googlesource.com/c/build/+/193357
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Chris Broadfoot <cbro@golang.org>