dashboard: skip testing x/net/http2/h2demo in GOPATH mode
The h2demo command is an internal command that we use to
powers the https://http2.golang.org server, it is not a
package that others import.

It was never tested before golang.org/issue/34361 because it
had a +build h2demo constraint. But that build constraint is
being removed, so explicitly skip testing it in GOPATH mode.

The package is supported only in module mode now, since
it requires third-party dependencies.

Updates golang/go#34361

Change-Id: Ia3870baa10688b83f8400b1a5e13824a1c1b3d3b
Reviewed-on: https://go-review.googlesource.com/c/build/+/196097
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>