app/appengine: fix up mistakes from not test deploying earlier CLs
During my earlier deletion spree over the past few CLs I wasn't
testing on real data (with production data) and didn't catch that
datastore returns an error when you load an old entity into a struct
that has since removed a field.

So filter those out.

(I could've also restored those fields, perhaps with a dummy named
interface{} type, but this code is temporary, and I'd prefer fewer
fields to think about. And this was a good opportunity to re-read the
code.)

I also still had references to the always-empty-string Prefix field in
the template.

Change-Id: I4fd9002d6d1bae2a08f4c53ec34651a7fe5db344
Reviewed-on: https://go-review.googlesource.com/c/build/+/208399
Reviewed-by: Bryan C. Mills <bcmills@google.com>