app/appengine: remove unused tagHandler, document others, remove auth from init
tagHandler isn't used by anything. Delete. Less code to update.

And document who calls the other handlers.

Also, remove the auth requirement from the /init handler. It's
harmless to call without auth, and the login restriction went away
with the Go 1.12+ runtime anyway, so deleting it gets us closer to
being able to use the Go 1.12/Go 1.13 runtimes. (The plan is to delete
most the code, port the small remaining bit to the cloud.google.com/go
libraries so we can update to Go 1.12/Go 1.13+, and then at that
point, since the cloud.google.com/go code will run anywhere, we can
just run it in the same process as the coordinator.)

Updates golang/go#34744

Change-Id: I929c70945a3e9e27b38b1d5899c7860470361927
Reviewed-on: https://go-review.googlesource.com/c/build/+/208678
Reviewed-by: Bryan C. Mills <bcmills@google.com>