dashboard: remove darwin-amd64-10_8 builder type
macOS 10.8 is very old and not supported by Go 1.13 nor Go 1.12.
There are no instances of this builder type available because
macOS builders are not created on demand and the ExpectNum value
is set to zero. Remove it and various files related to it.

Fixes golang/go#36560

Change-Id: Ife36624f3695fccff2f5a12af4518e25645b4f0d
Reviewed-on: https://go-review.googlesource.com/c/build/+/215257
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Run-TryBot: Brad Fitzpatrick <bradfitz@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>