maintner: handle deleted issues in githubRepoPoller.syncIssues
For an issue that does not exist, GitHub API returns the message
"Not Found", or when an issue is deleted, "This issue was deleted".

From https://developer.github.com/v3/issues/#get-a-single-issue:

> If the issue was transferred to or deleted from a repository
> where the authenticated user lacks read access, the API
> returns a 404 Not Found status. If the issue was deleted
> from a repository where the authenticated user has read
> access, the API returns a 410 Gone status.

Start checking ge.Response.StatusCode for http.StatusNotFound
or http.StatusGone rather than using the Message field, because
it is more robust and a closer match to GitHub API documentation.

An example GitHub API request which returns 410 Gone status is
https://api.github.com/repos/sourcegraph/sourcegraph/issues/2921.

Updates golang/go#37603

Change-Id: Ifb77fed722d720ecbeae9b5f196f200f8097a67f
Reviewed-on: https://go-review.googlesource.com/c/build/+/216438
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>