devapp/owners: remove mvdan from go/src/cmd/vet
The bulk of the source was refactored years ago with go/analysis and
moved into x/tools, and I haven't done any work to any parts of it
since.

I'm still interested in tools in general, but I'm not an owner of any of
it anymore. I don't have plans to pick that up again, for now.

Change-Id: Icbabeb5a091be816b90693c135af682a7910fa3a
Reviewed-on: https://go-review.googlesource.com/c/build/+/218458
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>