dashboard: treat dev branches as current
If a dev branch is active, it will be current. If it is not
active, it is probably not current, but we don't build on the
branch so it doesn't matter anyway.

Specifically, the dev.link branch is currently under active
development, and it would be good to have a large set of builders
run on the branch to catch problems. (E.g., AIX build was broken
but not noticed.)

Fixes golang/go#37953.

Change-Id: Idd7e3edaaff1b4d2cee17da552a96df2b00716e9
Reviewed-on: https://go-review.googlesource.com/c/build/+/223381
Run-TryBot: Cherry Zhang <cherryyz@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>