dashboard: use race-enabled linker in racecompile builder
When the compiler's backend became concurrent, we introduced the
racecompile builder to run the compiler with the race detector
enabled to catch races. In the new linker work the linker will
become more concurrent as well. It is probably a good idea to run
the linker with the race detector as well.

Change-Id: Ib50f030cba0509730714b74462f5a8942fe2fd2d
Reviewed-on: https://go-review.googlesource.com/c/build/+/226369
Run-TryBot: Cherry Zhang <cherryyz@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>