dashboard, app/appengine: set a known issue for staticlockranking builder
The linux-amd64-staticlockranking builder was added pre-emptively so
that it could be tested and used during the development of a CL that
implements a new feature. It's expected to fail until that work is
completed. Mark it as a builder with a known issue, golang/go#38029.

The new BuildConfig.KnownIssue field can be used by builders being
added in the future so that if they fail at first, it doesn't take
away time for people looking at build.golang.org, which we always
aim to have as green as possible¹.

¹ https://groups.google.com/d/msg/golang-dev/y0yM_Xi665Q/hUpBiUPiCAAJ

Fixes golang/go#38283
For golang/go#38029
For golang/go#37937
For golang/go#11811

Change-Id: Iba1606c7021ffa7e67ddbaae2fc6b65f4e46ab34
Reviewed-on: https://go-review.googlesource.com/c/build/+/227379
Reviewed-by: Alexander Rakoczy <alex@golang.org>