delete update-deps.sh
This script was used when a custom gitlock tool was used for package
management. That approach has been superseded with the switch to the
module mode, and gitlock was deleted in CL 176604. By now, there are
no directories with an update-deps make target remaining.

For golang/go#26872.

Change-Id: Ia7f5d9e8b277be55010d5752523d1160b72115b2
Reviewed-on: https://go-review.googlesource.com/c/build/+/227440
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Run-TryBot: Alexander Rakoczy <alex@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>