dashboard, env/netbsd-{amd64,386}: mount /tmp on disk instead of RAM
The /tmp partition was being mounted as tmpfs -s=ram%25 by default.
On the n1-highcpu-4 machine type with 3.6 GB RAM, this resulted in
there being less than 1 GB of available disk space, which is not
sufficient to build Go and run its tests.

CL 226977 has increased the image disk size from 4 GB to 16 GB, so
using the disk for /tmp resolves issues with insufficient disk space.

I also considered using an alternative machine type like n1-standard-4
with 15 GB RAM instead, but didn't pursue that option for now because
it was not clear if it could be made to work for the 32-bit builder.

Remove the replacement of wd0 with sd0 in /etc/fstab. It is no longer
required because the NetBSD 9.0 installer defaults to installing on GPT
rather than MBR as was done in earlier versions.

For golang/go#37244.
For golang/go#30839.
For golang/go#31726.

Change-Id: Icbd0f909ab7768d775c96a6b606b5279a600f1bd
Reviewed-on: https://go-review.googlesource.com/c/build/+/227538
Reviewed-by: Benny Siegert <bsiegert@gmail.com>
Run-TryBot: Benny Siegert <bsiegert@gmail.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>